Refactor agent structure and update imports
- Moved customer service agent functionality to a new `banking_agent.py` file, updating relevant imports across the codebase.
- Introduced new agents for caller and insurance functionalities, enhancing modularity and organization.
- Updated bridge files to reflect changes in agent registration and session configuration.
- Adjusted test files to align with the new agent structure, ensuring proper function registration and testing for customer service interactions.
