+++ opusagent/main.py
@@ -29,7 +29,7 @@ from opusagent.config import get_config, server_config, mock_config, vad_config,
 from opusagent.config.env_loader import load_env_file
 from opusagent.config.logging_config import configure_logging
 from opusagent.config.models import WebSocketConfig
-from opusagent.customer_service_agent import session_config
+from opusagent.agents.banking_agent import session_config
 from opusagent.session_manager import SessionManager
 from opusagent.websocket_manager import get_websocket_manager, WebSocketManager
 from opusagent.callers import get_available_caller_types, get_caller_description
